feat: audit history refactor
